add classe combate e alterando variavel raça para  poder do mal ao vilao

+++ src/Guerreiro.java
@@ -10,7 +10,7 @@ public class Guerreiro extends Personagem {
         this.bonusVida = bonusVida;
     }
 
-    public float getBonusVida() {
+    public int getBonusVida() {
         return bonusVida;
     }
